chore: moved icon components
