+++ fableous-fe/src/components/canvas/BrushWidthIcon.tsx
