+++ fableous-fe/src/components/canvas/EraserIcon.tsx
